update num_and to uint64_t

+++ emp-tool/execution/circuit_execution.h
@@ -19,7 +19,7 @@ class CircuitExecution { public:
 	virtual block xor_gate(const block&in1, const block&in2) = 0;
 	virtual block not_gate(const block& in1) = 0;
 	virtual block public_label(bool b) = 0;
-	virtual size_t num_and() {
+	virtual uint64_t num_and() {
 		return -1;
 	}
 	virtual ~CircuitExecution (){ }
