+++ emp-tool/execution/plain_circ.h
@@ -131,7 +131,7 @@ public:
 			return res;
 		}
 	}
-	size_t num_and() override {
+	uint64_t num_and() override {
 		return ands;
 	}
