+++ emp-tool/gc/privacy_free_gen.h
@@ -78,7 +78,7 @@ public:
 	block not_gate(const block& a) {
 		return a ^ public_label(true);
 	}
-	size_t num_and() override {
+	uint64_t num_and() override {
 		return gid;
 	}
 };
